+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Implement calculated properties for Performance objects (from @justjasongreen)
+- Implement performance lists via the PerformanceList class (from @justjasongreen)
 
 
 ## [1.0.0b2] - 2016-07-24